refactor(core): remove 2 @deprecated type re-export shims in core/video/
删除 2 个标 @deprecated 但实际为向后兼容 re-export 的文件：

- src/core/video/types.ts (16 行)
- src/core/video/highlight-types.ts (13 行)

迁移 10 个 internal consumer 从 './types' / 'highlight-types' 改用 '@/types'：
- core/video/{ivideo-processor,base-video-processor,tauri-video-processor,index}.ts
- core/services/ai-clip/{analyzer,segment-enricher,types}.ts
- core/services/video/transition-suggestion.ts (+ .test.ts)
- components/ai-clip/components/segment-timeline-card.tsx

-11 / -41 行（net）。
公开类型路径不变（仍可从 @/types 导入）。
75/75 测试全绿，type-check 干净。

注：spec 列出的 5 项清理中：
- 2 个 vision BETA services（scene/object-detection）实际被新 vision/index.ts 使用，
  非零引用，跳过（需要单独 PR 替换 Math.random 为真 Rust tauri 调用）。
- useStoryFabStore / ProjectStatus 各有 1 / 4 个 consumer，需先迁移后删除，单独 PR 处理。

Stage 8 PR-1.4 / spec 阶段 1。
