--- src/core/video/highlight-types.ts
@@ -1,13 +1,0 @@
-/**
- * @deprecated 请从 @/types 导入高光检测相关类型
- * 此文件为向后兼容的重导出层，将在后续版本移除
- */
-export type {
-  HighlightSegment,
-  HighlightOptions,
-  DetectHighlightsInput,
-  SegmentType,
-  SmartVideoSegment,
-  SegmentOptions,
-  DetectSmartSegmentsInput,
-} from '@/types';
