+++ src/core/video/index.ts
@@ -2,11 +2,12 @@
  * Video Processing Module
  *
  * 架构：
- * - types.ts              → 核心类型定义
  * - IVideoProcessor.ts    → 接口（后端无关）
  * - BaseVideoProcessor.ts → 基类（通用逻辑：错误归一化、FFmpeg 缓存）
  * - TauriVideoProcessor.ts → Tauri invoke 实现
  * - formatters.ts         → 纯格式化函数
+ *
+ * 类型已统一从 @/types 导入（2026-07 Stage 8 PR-1.4）
  *
  * 使用方式：
  *   import { videoProcessor } from '@/core/video';
@@ -18,8 +19,6 @@
  *     ...
  *   }
  */
-export * from './types';
-export * from './highlight-types';
 export * from './ivideo-processor';
 export { BaseVideoProcessor, VideoProcessingError, normalizeVideoError } from './base-video-processor';
 export { videoProcessor, TauriVideoProcessor } from './tauri-video-processor';
