--- src/core/video/types.ts
@@ -1,16 +1,0 @@
-/**
- * @deprecated 请从 @/types 导入视频处理相关类型
- * 此文件为向后兼容的重导出层，将在后续版本移除
- */
-export type {
-  VideoMetadata,
-  KeyFrame,
-  VideoSegment,
-  SimpleVideoSegment,
-  ProcessingProgress,
-  ProgressCallback,
-  ExtractKeyFramesOptions,
-  TranscodeOptions,
-  CutOptions,
-  FFmpegStatus,
-} from '@/types';
